# docs/openapi/openapi.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"LegacyRestrictionDetailsResponse": {
				"properties": {
					"response": {
						"properties": {
							"errorMsg": {
								"type": "string"
							},
							"errorCode": {
								"type": "string"
							},
							"respMsg": {
								"type": "string"
							},
							"respCodeMsg": {
								"type": "string"
							},
							"respCode": {
								"type": "string"
							},
							"respStatus": {
								"type": "boolean"
							}
						},
						"required": [
							"errorMsg",
							"errorCode",
							"respMsg",
							"respCodeMsg",
							"respCode",
							"respStatus"
						],
						"type": "object"
					},
					"data": {
						"$ref": "#/components/schemas/Record_string.unknown_"
					}
				},
				"required": [
					"response",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"LegacyRestrictionDeniedResponse": {
				"properties": {
					"result": {
						"type": "boolean"
					},
					"restrictions": {
						"$ref": "#/components/schemas/Record_string.boolean_"
					}
				},
				"required": [
					"result",
					"restrictions"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LegacyUserRestrictedRequest": {
				"properties": {
					"customer": {
						"type": "number",
						"format": "double"
					},
					"user": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"customer",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LegacyUserFwCloudAccessRequest": {
				"properties": {
					"user": {
						"type": "number",
						"format": "double"
					},
					"fwcloud": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"user",
					"fwcloud"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LegacyFwCloudAccessItem": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"created_at": {
						"type": "string"
					},
					"updated_at": {
						"type": "string"
					},
					"created_by": {
						"type": "number",
						"format": "double"
					},
					"updated_by": {
						"type": "number",
						"format": "double"
					},
					"locked_at": {
						"type": "string",
						"nullable": true
					},
					"locked_by": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"locked": {
						"type": "number",
						"format": "double"
					},
					"image": {
						"type": "string"
					},
					"comment": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"created_at",
					"updated_at",
					"created_by",
					"updated_by",
					"locked_at",
					"locked_by",
					"locked",
					"image",
					"comment"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LegacyUserFwCloudAccessListRequest": {
				"properties": {
					"user": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LegacyFwcloudResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"created_at": {
						"type": "string"
					},
					"updated_at": {
						"type": "string"
					},
					"created_by": {
						"type": "number",
						"format": "double"
					},
					"updated_by": {
						"type": "number",
						"format": "double"
					},
					"locked_at": {
						"type": "string",
						"nullable": true
					},
					"locked_by": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"locked": {
						"type": "number",
						"format": "double"
					},
					"image": {
						"type": "string"
					},
					"comment": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"created_at",
					"updated_at",
					"created_by",
					"updated_by",
					"locked_at",
					"locked_by",
					"locked",
					"image",
					"comment"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LegacyFwcloudErrorResponse": {
				"properties": {
					"fwcErr": {
						"type": "number",
						"format": "double"
					},
					"msg": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"LegacyFwcloudRequest": {
				"properties": {
					"fwcloud": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"fwcloud"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"LegacyFwcloudLockResponse": {
				"properties": {
					"result": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"info": {
						"$ref": "#/components/schemas/LegacyFwcloudLockInfo"
					}
				},
				"required": [
					"result",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"LegacyFirewallErrorResponse": {
				"properties": {
					"fwcErr": {
						"type": "number",
						"format": "double"
					},
					"msg": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"LegacyFirewallResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"cluster": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"fwcloud": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"comment": {
						"type": "string",
						"nullable": true
					},
					"created_at": {
						"type": "string"
					},
					"updated_at": {
						"type": "string"
					},
					"compiled_at": {
						"type": "string",
						"nullable": true
					},
					"installed_at": {
						"type": "string",
						"nullable": true
					},
					"by_user": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"install_user": {
						"type": "string"
					},
					"install_pass": {
						"type": "string"
					},
					"save_user_pass": {
						"type": "number",
						"format": "double"
					},
					"install_interface": {
						"type": "string",
						"nullable": true
					},
					"install_ipobj": {
						"type": "string",
						"nullable": true
					},
					"fwmaster": {
						"type": "number",
						"format": "double"
					},
					"install_port": {
						"type": "number",
						"format": "double"
					},
					"options": {
						"type": "number",
						"format": "double"
					},
					"interface_name": {
						"type": "string",
						"nullable": true
					},
					"ip_name": {
						"type": "string",
						"nullable": true
					},
					"ip": {
						"type": "string",
						"nullable": true
					},
					"id_fwmaster": {
						"type": "number",
						"format": "double",
						"nullable": true
					}
				},
				"required": [
					"id",
					"cluster",
					"fwcloud",
					"name",
					"comment",
					"created_at",
					"updated_at",
					"compiled_at",
					"installed_at",
					"by_user",
					"status",
					"install_user",
					"install_pass",
					"save_user_pass",
					"install_interface",
					"install_ipobj",
					"fwmaster",
					"install_port",
					"options",
					"interface_name",
					"ip_name",
					"ip",
					"id_fwmaster"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"LegacyFirewallCloudRequest": {
				"properties": {
					"fwcloud": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"fwcloud"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"LegacyCustomerRequest": {
				"properties": {
					"customer": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"customer"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LegacyCustomerRestrictionResponse": {
				"properties": {
					"result": {
						"type": "boolean"
					},
					"restrictions": {
						"$ref": "#/components/schemas/Record_string.boolean_"
					}
				},
				"required": [
					"result",
					"restrictions"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"LegacyClusterErrorResponse": {
				"properties": {
					"fwcErr": {
						"type": "number",
						"format": "double"
					},
					"msg": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"LegacyClusterGetByCloudRequest": {
				"properties": {
					"fwcloud": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"fwcloud"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "FWCloud API",
		"version": "2.4.1",
		"description": "FWCloud RESTful API.",
		"contact": {}
	},
	"paths": {
...
		"/user/restricted": {
			"put": {
				"operationId": "Restricted",
				"responses": {
					"200": {
						"description": "Restriction details",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyRestrictionDetailsResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"response": {
												"respStatus": true,
												"respCode": "ACR_OK",
												"respCodeMsg": "Ok",
												"respMsg": "",
												"errorCode": "",
												"errorMsg": ""
											},
											"data": {}
										}
									}
								}
							}
						}
					},
					"204": {
						"description": "No deletion restrictions"
					},
					"403": {
						"description": "Legacy endpoint error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyRestrictionDeniedResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"result": true,
											"restrictions": {
												"CustomerHasUsers": true
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Check that there are no restrictions for user deletion.",
				"summary": "Restrictions for user deletion.",
				"tags": [
					"user"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LegacyUserRestrictedRequest"
							},
							"example": {
								"customer": 10,
								"user": 5
							}
						}
					}
				}
			}
		},
...
		"/user/fwcloud/del": {
			"put": {
				"operationId": "DisableFwCloudAccess",
				"responses": {
					"204": {
						"description": "Disable cloud access"
					}
				},
				"description": "Disable user access to a fwcloud.",
				"summary": "Disable cloud access.",
				"tags": [
					"user"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LegacyUserFwCloudAccessRequest"
							},
							"example": {
								"user": 5,
								"fwcloud": 2
							}
						}
					}
				}
			}
		},
		"/user/fwcloud/get": {
			"put": {
				"operationId": "ListFwCloudAccess",
				"responses": {
					"200": {
						"description": "List of fwclouds with access",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/LegacyFwCloudAccessItem"
									},
									"type": "array"
								},
								"examples": {
									"Example 1": {
										"value": [
											{
												"id": 4,
												"name": "FWCloud-02",
												"created_at": "2019-05-14T11:37:19.000Z",
												"updated_at": "2019-05-14T11:37:19.000Z",
												"created_by": 0,
												"updated_by": 0,
												"locked_at": null,
												"locked_by": null,
												"locked": 0,
												"image": "",
												"comment": ""
											},
											{
												"id": 5,
												"name": "FWCloud-03",
												"created_at": "2019-05-14T11:37:24.000Z",
												"updated_at": "2019-05-14T11:57:06.000Z",
												"created_by": 0,
												"updated_by": 0,
												"locked_at": "2019-05-14T11:57:06.000Z",
												"locked_by": 1,
												"locked": 1,
												"image": "",
												"comment": ""
											}
										]
									}
								}
							}
						}
					}
				},
				"description": "List of fwclouds to which the indicated user has access to.",
				"summary": "List of fwclouds with access.",
				"tags": [
					"user"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LegacyUserFwCloudAccessListRequest"
							},
							"example": {
								"user": 5
							}
						}
					}
				}
			}
		},
		"/fwcloud/all/get": {
			"get": {
				"operationId": "GetAllowed",
				"responses": {
					"200": {
						"description": "Get allowed fwclouds",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/LegacyFwcloudResponse"
									},
									"type": "array"
								},
								"examples": {
									"Example 1": {
										"value": [
											{
												"id": 4,
												"name": "FWCloud-02",
												"created_at": "2019-05-14T11:37:19.000Z",
												"updated_at": "2019-05-14T11:37:19.000Z",
												"created_by": 0,
												"updated_by": 0,
												"locked_at": null,
												"locked_by": null,
												"locked": 0,
												"image": "",
												"comment": ""
											},
											{
												"id": 5,
												"name": "FWCloud-03",
												"created_at": "2019-05-14T11:37:24.000Z",
												"updated_at": "2019-05-14T11:57:06.000Z",
												"created_by": 0,
												"updated_by": 0,
												"locked_at": "2019-05-14T11:57:06.000Z",
												"locked_by": 1,
												"locked": 1,
												"image": "",
												"comment": ""
											}
										]
									}
								}
							}
						}
					},
					"204": {
						"description": "No allowed fwclouds"
					},
					"400": {
						"description": "Legacy endpoint error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyFwcloudErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"fwcErr": 7000,
											"msg": "FWCloud access not allowed"
										}
									}
								}
							}
						}
					}
				},
				"description": "Get fwcloud data for all the fwclouds to which the logged user has access.",
				"summary": "Get allowed fwclouds.",
				"tags": [
					"fwcloud"
				],
				"security": [],
				"parameters": []
			}
		},
		"/fwcloud/get": {
			"put": {
				"operationId": "Get",
				"responses": {
					"200": {
						"description": "Get fwcloud data",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyFwcloudResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"id": 3,
											"name": "FWCloud-Updated",
											"created_at": "2019-05-14T11:37:15.000Z",
											"updated_at": "2019-05-14T11:37:54.000Z",
											"created_by": 0,
											"updated_by": 0,
											"locked_at": "2019-05-14T11:37:51.000Z",
											"locked_by": 1,
											"locked": 1,
											"image": "",
											"comment": "Comment for the updated fwcloud."
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Legacy endpoint error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyFwcloudErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"fwcErr": 7000,
											"msg": "FWCloud access not allowed"
										}
									}
								}
							}
						}
					}
				},
				"description": "Get fwcloud data.",
				"summary": "Get fwcloud data.",
				"tags": [
					"fwcloud"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LegacyFwcloudRequest"
							},
							"example": {
								"fwcloud": 3
							}
						}
					}
				}
			}
		},
		"/fwcloud/restricted": {
			"put": {
				"operationId": "Restricted",
				"responses": {
					"204": {
						"description": "No deletion restrictions"
					},
					"400": {
						"description": "Legacy endpoint error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyFwcloudErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"fwcErr": 7000,
											"msg": "FWCloud access not allowed"
										}
									}
								}
							}
						}
					}
				},
				"description": "Check if the fwcloud indicated as a parameter has any deletion restriction.",
				"summary": "Check delete restrictions.",
				"tags": [
					"fwcloud"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LegacyFwcloudRequest"
							},
							"example": {
								"fwcloud": 2
							}
						}
					}
				}
			}
		},
		"/fwcloud/del": {
			"put": {
				"operationId": "Delete",
				"responses": {
					"204": {
						"description": "Delete fwcloud"
					},
					"400": {
						"description": "Legacy endpoint error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyFwcloudErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"fwcErr": 7000,
											"msg": "FWCloud access not allowed"
										}
									}
								}
							}
						}
					}
				},
				"description": "Delete the firewall cloud indicated in the request body.",
				"summary": "Delete fwcloud.",
				"tags": [
					"fwcloud"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LegacyFwcloudRequest"
							},
							"example": {
								"fwcloud": 2
							}
						}
					}
				}
			}
		},
		"/fwcloud/lock": {
			"put": {
				"operationId": "Lock",
				"responses": {
					"200": {
						"description": "Lock result",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyFwcloudLockResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"result": true,
											"message": "FWCLOUD LOCKED OK"
										}
									}
								}
							}
						}
					}
				},
				"description": "Lock fwcloud status.",
				"summary": "Lock fwcloud.",
				"tags": [
					"fwcloud"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LegacyFwcloudRequest"
							},
							"example": {
								"fwcloud": 2
							}
						}
					}
				}
			}
		},
		"/fwcloud/unlock": {
			"put": {
				"operationId": "Unlock",
				"responses": {
					"200": {
						"description": "Unlock result",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyFwcloudLockResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"result": true,
											"message": "FWCLOUD UNLOCKED OK"
										}
									}
								}
							}
						}
					}
				},
				"description": "Unlock fwcloud status.",
				"summary": "Unlock fwcloud.",
				"tags": [
					"fwcloud"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LegacyFwcloudRequest"
							},
							"example": {
								"fwcloud": 2
							}
						}
					}
				}
			}
		},
		"/fwcloud/forcelock": {
			"put": {
				"operationId": "ForceUnlock",
				"responses": {
					"200": {
						"description": "Force unlock result",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyFwcloudLockResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"result": true,
											"message": "FWCLOUD FORCE UNLOCKED OK"
										}
									}
								}
							}
						}
					}
				},
				"description": "Force unlock fwcloud status.",
				"summary": "Force unlock fwcloud.",
				"tags": [
					"fwcloud"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LegacyFwcloudRequest"
							},
							"example": {
								"fwcloud": 2
							}
						}
					}
				}
			}
		},
...
		"/firewall/cloud/get": {
			"put": {
				"operationId": "GetByCloud",
				"responses": {
					"200": {
						"description": "Get firewalls in cloud",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/LegacyFirewallResponse"
									},
									"type": "array"
								},
								"examples": {
									"Example 1": {
										"value": [
											{
												"id": 1,
												"cluster": null,
												"fwcloud": 1,
												"name": "Firewall-01",
												"comment": null,
												"created_at": "2019-05-15T10:34:46.000Z",
												"updated_at": "2019-05-15T10:34:47.000Z",
												"compiled_at": null,
												"installed_at": null,
												"by_user": 1,
												"status": 3,
												"install_user": "",
												"install_pass": "",
												"save_user_pass": 0,
												"install_interface": null,
												"install_ipobj": null,
												"fwmaster": 0,
												"install_port": 22,
												"options": 0,
												"interface_name": null,
												"ip_name": null,
												"ip": null,
												"id_fwmaster": null
											},
											{
												"id": 2,
												"cluster": null,
												"fwcloud": 1,
												"name": "Firewall-02",
												"comment": null,
												"created_at": "2019-05-15T10:34:46.000Z",
												"updated_at": "2019-05-15T10:34:47.000Z",
												"compiled_at": null,
												"installed_at": null,
												"by_user": 1,
												"status": 3,
												"install_user": "",
												"install_pass": "",
												"save_user_pass": 0,
												"install_interface": null,
												"install_ipobj": null,
												"fwmaster": 0,
												"install_port": 22,
												"options": 0,
												"interface_name": null,
												"ip_name": null,
												"ip": null,
												"id_fwmaster": null
											}
										]
									}
								}
							}
						}
					},
					"400": {
						"description": "Legacy endpoint error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyFirewallErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"fwcErr": 7001,
											"msg": "Firewall access not allowed"
										}
									}
								}
							}
						}
					}
				},
				"description": "Get firewalls data by fwcloud.",
				"summary": "Get firewalls in cloud.",
				"tags": [
					"firewall"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LegacyFirewallCloudRequest"
							},
							"example": {
								"fwcloud": 1
							}
						}
					}
				}
			}
		},
...
		"/customer/restricted": {
			"put": {
				"operationId": "Restricted",
				"responses": {
					"204": {
						"description": "No deletion restrictions"
					},
					"403": {
						"description": "Legacy endpoint restriction",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyCustomerRestrictionResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"result": true,
											"restrictions": {
												"CustomerHasUsers": true
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Check that there are no restrictions for customer deletion.",
				"summary": "Restrictions for customer deletion.",
				"tags": [
					"customer"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LegacyCustomerRequest"
							},
							"example": {
								"customer": 10
							}
						}
					}
				}
			}
		},
...
		"/cluster/cloud/get": {
			"put": {
				"operationId": "GetByCloud",
				"responses": {
					"200": {
						"description": "Get cloud clusters",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/LegacyClusterSummary"
									},
									"type": "array"
								},
								"examples": {
									"Example 1": {
										"value": [
											{
												"id": 1,
												"fwcloud": 2,
												"name": "Cluster-01",
												"comment": null,
												"created_at": "2019-05-17T11:46:57.000Z",
												"updated_at": "2019-05-17T11:46:57.000Z",
												"created_by": 0,
												"updated_by": 0
											},
											{
												"id": 2,
												"fwcloud": 2,
												"name": "Cluster-02",
												"comment": null,
												"created_at": "2019-05-17T11:47:00.000Z",
												"updated_at": "2019-05-17T11:47:00.000Z",
												"created_by": 0,
												"updated_by": 0
											}
										]
									}
								}
							}
						}
					},
					"400": {
						"description": "Legacy endpoint error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyClusterErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"fwcErr": 7002,
											"msg": "Cluster access not allowed"
										}
									}
								}
							}
						}
					}
				},
				"description": "Get all the cluster data for the indicated fwcloud.",
				"summary": "Get cloud clusters.",
				"tags": [
					"cluster"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LegacyClusterGetByCloudRequest"
							},
							"example": {
								"fwcloud": 2
							}
						}
					}
				}
			}
		},
...
		"/cluster/restricted": {
			"put": {
				"operationId": "Restricted",
				"responses": {
					"204": {
						"description": "No deletion restrictions"
					},
					"403": {
						"description": "Legacy endpoint restriction",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyClusterRestrictionsResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"result": true,
											"restrictions": {
												"FirewallHasInterfaces": true
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Check if there are restrictions for cluster deletion.",
				"summary": "Restrictions for cluster deletion.",
				"tags": [
					"cluster"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LegacyClusterGetRequest"
							},
							"example": {
								"fwcloud": 2,
								"cluster": 5
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	],
...
}